+++ example/drop/game.dart
@@ -3,6 +3,7 @@ library drop;
 import 'dart:html';
 import 'package:boarding/pieces.dart';
 import 'package:boarding/boarding.dart';
+import 'package:boarding/util.dart';
 
 part 'view/board.dart';